+++ src/Data/TypedEncoding/Internal/Class.hs
@@ -16,9 +16,10 @@ module Data.TypedEncoding.Internal.Class (
     , module Data.TypedEncoding.Internal.Class.Util
     , module Data.TypedEncoding.Internal.Class.Encode
     , module Data.TypedEncoding.Internal.Class.Decode
+    , module Data.TypedEncoding.Internal.Class.Validate 
     , module Data.TypedEncoding.Internal.Deprecated.Encode
     , module Data.TypedEncoding.Internal.Deprecated.Decode
-    , module Data.TypedEncoding.Internal.Class.Recreate 
+    , module Data.TypedEncoding.Internal.Deprecated.Recreate 
     , module Data.TypedEncoding.Internal.Class.Superset 
     -- * Encoder and Encoding replace EncodeFAll
     , module Data.TypedEncoding.Internal.Class.Encoder 
@@ -27,9 +28,10 @@ module Data.TypedEncoding.Internal.Class (
 import           Data.TypedEncoding.Internal.Class.Util
 import           Data.TypedEncoding.Internal.Class.Encode
 import           Data.TypedEncoding.Internal.Class.Decode
+import           Data.TypedEncoding.Internal.Class.Validate
 import           Data.TypedEncoding.Internal.Deprecated.Encode
 import           Data.TypedEncoding.Internal.Deprecated.Decode
-import           Data.TypedEncoding.Internal.Class.Recreate
+import           Data.TypedEncoding.Internal.Deprecated.Recreate
 import           Data.TypedEncoding.Internal.Class.Superset
 import           Data.TypedEncoding.Internal.Class.Encoder