+++ src/Data/TypedEncoding/Internal/Types/Enc.hs
@@ -67,8 +67,15 @@ unsafeSetPayload  = MkEnc Proxy
 withUnsafeCoerce ::  (s1 -> s2) -> Enc e1 c s1 -> Enc e2 c s2
 withUnsafeCoerce f (MkEnc _ conf str)  = MkEnc Proxy conf (f str)
 
+withUnsafeCoerceF :: forall e1 e2 f c s1 s2 . Functor f => (s1 -> f s2) -> Enc e1 c s1 -> f (Enc e2 c s2)
+withUnsafeCoerceF f (MkEnc _ conf str)  = MkEnc Proxy conf <$> f str
+
 unsafeChangePayload ::  (s1 -> s2) -> Enc e c s1 -> Enc e c s2
 unsafeChangePayload f (MkEnc p conf str)  = MkEnc p conf (f str) 
+
+getTransformF :: forall e1 e2 f c s1 s2 . Functor f => (Enc e1 c s1 -> f (Enc e2 c s2)) -> c -> s1 -> f s2
+getTransformF fn c str = getPayload <$> fn (unsafeSetPayload c str)
+ 
 
 -- |
 -- Wraps encoding function that adds encoding to existing stack.